# Increase SCURVE cycle count for unoptimized cortex-m0