--- a/Marlin/src/module/stepper.h
+++ b/Marlin/src/module/stepper.h
@@ -76,7 +76,7 @@
   #define TIMER_READ_ADD_AND_STORE_CYCLES 34UL
 
   // The base ISR takes 792 cycles
-  #if defined(STM32G0B1xx)
+  #ifdef STM32G0B1xx
     #define ISR_BASE_CYCLES  928UL
   #else
     #define ISR_BASE_CYCLES  792UL
@@ -84,7 +84,7 @@
 
   // Linear advance base time is 64 cycles
   #if ENABLED(LIN_ADVANCE)
-    #if defined(STM32G0B1xx)
+    #ifdef STM32G0B1xx
       #define ISR_LA_BASE_CYCLES 200UL
     #else
       #define ISR_LA_BASE_CYCLES 64UL
@@ -95,7 +95,7 @@
 
   // S curve interpolation adds 40 cycles
   #if ENABLED(S_CURVE_ACCELERATION)
-    #if defined(STM32G0B1xx)
+    #ifdef STM32G0B1xx
       #define ISR_S_CURVE_CYCLES 544UL
     #else
       #define ISR_S_CURVE_CYCLES 40UL