@cbagwell 2022-11-09
Further testing in #24927 showed `eval_bezier_curve()` is closer to 544 than 500 so updating that value.

`Stepper::calc_timer_interval()` was also identified as taking a long time since its division is translating to `_aeabi_uidiv()` at 136 cycles (compared to ARM CPU's with udiv instruction taking 12).  This seems high enough it should be accounted for.

Its called up to 2 times on accel/decel/coast in `Stepper::block_phase_isr()` (one is only if LIN_ADVANCE is enabled).

I'm not sure how best to account for that so I put one call in `IS_BASE_CYCLES` and one in `ISR_LA_BASE_CYCLES`.

I'm purposely ignoring input shaping high runners since its still in active development.