@thinkyhead 2022-11-11
Thank you for your time and attention at this level of the system. It's always the trickiest area to optimize and rework while keeping everything else stable.

Always looking for new ways to optimize, if we can eliminate the division by a pre-calculation of some reciprocal in the main thread that would be one possible solution. Another would be to look at the maths leading up to the division and see if a reciprocal or divisor can be figured earlier, and/or stored in a static and only updated as needed.

It sounds like this should be merged anyway, as some Input Shaping work will depend on it, so let me give a quick review and cleanup, and hopefully this is in ready form….