@cbagwell 2022-11-11
Thanks for reviewing.

A quick update that is new in last 12 hours or so.  @tombrazier just pushed an update to #24951 and this PR will now have a merge conflict with that PR.

The good news is that his PR improves the ISR's to now allow the combination of INPUT_SHAPING+ADAPTIVE_STEP_SMOOTHING+S_CURVE_ACCELERATION to work on STM32G0's even with under estimated SCRUVE cycles.  I can guess it might help BLTouch users as well.

That PR update is so new I was still debating what to do with this PR.  Although this PR does allow INPUT_SHAPING+ADAPTIVE_STEP_SMOOTHING+S_CURVE_ACCELERATION as well, I'd obviously pick #24951 over this PR if all things equal.

It's probably worth reducing this PR back down to updating the SCURVE cycles only but at same time @tombrazier also has a prototype of optimizing SCURVE for STM32G0's that will hopefully reduce the 544 cycle count alot.