@tombrazier 2022-11-11
> Further testing in https://github.com/MarlinFirmware/Marlin/issues/24927 showed eval_bezier_curve() is closer to 544 than 500 so updating that value.

I am now not certain I measured that 544 correctly. I would trust your examination of the disassembled code more.

> Always looking for new ways to optimize, if we can eliminate the division by a pre-calculation of some reciprocal in the main thread that would be one possible solution. Another would be to look at the maths leading up to the division and see if a reciprocal or divisor can be figured earlier, and/or stored in a static and only updated as needed.

The division is to get time to the next step from step rate. No way round it. But for the AVRs it is replaced with a lookup table. The same could be done for ARMs without UDIV. But it would need some work.